+++ CHANGES.md
@@ -1,7 +1,8 @@
 v0.9.0 2015-06-17 Cambridge (UK)
 --------------------------------
 
-Support for Unicode 8.0.0's new line breaking and sentence boundary rules.
+- Support for Unicode 8.0.0's new line breaking and sentence boundary rules.
+- `Uuseg.custom` add a unit argument.
 
 
 v0.8.0 2014-12-23 Cugy (VD)