+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+- Build depend on topkg.
 - Relicense from BSD3 to ISC.
 
 v0.9.0 2015-06-17 Cambridge (UK)